learn

+++ src/ACM/ID1005.java
@@ -7,8 +7,8 @@ public class ID1005 {
     public static void main(String[] args) {
         int num = scan.nextInt();
         for(int i = 1;i<=num;i++){
-            float x = Float.valueOf(scan.nextFloat());
-            float y = Float.valueOf(scan.nextFloat());
+            float x = scan.nextFloat();
+            float y = scan.nextFloat();
             double r2 = x*x+y*y;
             int year = 1;
             int mj = 0;
